✨ feat(VSecM): 448 Remove unnecessary classes
