--- sdk-java/src/main/java/org/wsecm/task/Task.java
@@ -1,5 +1,0 @@
-package org.wsecm.task;
-
-public interface Task {
-    void execute();
-}
