--- sdk-java/src/main/java/org/wsecm/task/TaskFactory.java
@@ -1,8 +1,0 @@
-package org.wsecm.task;
-
-import java.util.concurrent.ScheduledExecutorService;
-
-public interface TaskFactory {
-    Task createTask(ScheduledExecutorService scheduler);
-}
-
