--- sdk-java/src/main/java/org/wsecm/task/backoff/BackoffStrategy.java
@@ -1,6 +1,0 @@
-package org.wsecm.task.backoff;
-
-public interface BackoffStrategy {
-    long calculateBackoff(long interval, long successCount, long errorCount);
-}
-
